add comment likes and post likes
