+++ src/Blog/Repositories/LikeRepository/SqliteCommentLikesRepository.php
@@ -3,16 +3,14 @@
 namespace Veliafar\PhpBlog\Blog\Repositories\LikeRepository;
 
 use PDO;
-use Veliafar\PhpBlog\Blog\CommentLike;
 use Veliafar\PhpBlog\Blog\Exceptions\CommentNotFoundException;
 use Veliafar\PhpBlog\Blog\Exceptions\InvalidArgumentException;
 use Veliafar\PhpBlog\Blog\Exceptions\LikeAlreadyExistException;
 use Veliafar\PhpBlog\Blog\Exceptions\LikeNotFoundException;
 use Veliafar\PhpBlog\Blog\Exceptions\PostNotFoundException;
 use Veliafar\PhpBlog\Blog\Exceptions\UserNotFoundException;
-use Veliafar\PhpBlog\Blog\PostLike;
+use Veliafar\PhpBlog\Blog\Like\CommentLike;
 use Veliafar\PhpBlog\Blog\Repositories\CommentRepository\SqliteCommentsRepository;
-use Veliafar\PhpBlog\Blog\Repositories\PostRepository\SqlitePostsRepository;
 use Veliafar\PhpBlog\Blog\Repositories\UsersRepository\SqliteUsersRepository;
 use Veliafar\PhpBlog\Blog\UUID;
