+++ src/Http/Actions/Like/CreatePostLike.php
@@ -7,7 +7,7 @@ use Veliafar\PhpBlog\Blog\Exceptions\InvalidArgumentException;
 use Veliafar\PhpBlog\Blog\Exceptions\LikeAlreadyExistException;
 use Veliafar\PhpBlog\Blog\Exceptions\PostNotFoundException;
 use Veliafar\PhpBlog\Blog\Exceptions\UserNotFoundException;
-use Veliafar\PhpBlog\Blog\PostLike;
+use Veliafar\PhpBlog\Blog\Like\PostLike;
 use Veliafar\PhpBlog\Blog\Repositories\LikeRepository\PostLikeRepositoryInterface;
 use Veliafar\PhpBlog\Blog\Repositories\PostRepository\PostRepositoryInterface;
 use Veliafar\PhpBlog\Blog\Repositories\UsersRepository\UserRepositoryInterface;
